api/requirements.txt
# flask
# flask-cors
# python-dotenv
# pinecone-client
# langchain
# langchain-openai
# langchain-pinecone
# openai
# werkzeug
# logging
flask>=2.3.3
flask-cors>=4.0.0
python-dotenv>=1.0.1
pinecone-client>=3.2.2
langchain==0.3.27
langchain-openai>=0.1.7
langchain-pinecone>=0.1.0
openai>=1.3.7
werkzeug>=2.3.7
flask-sqlalchemy>=3.0.0
flask-login>=0.6.3
bcrypt>=4.0.1
PyJWT>=2.8.0

# Compatibility pins for modern stack
pydantic>=2.7.4
typing_extensions>=4.8.0

# Keep a bare pinecone entry removed (use pinecone-client)
# pinecone
